+++ lib/src/models/wagon_wheel_batsman_properties.dart
@@ -39,4 +39,3 @@ class WagonWheelBatsmanProperties {
   String get effectiveIconPath =>
       iconPath ?? WagonWheelPackageConstants.defaultBatsmanIconPath;
 }
-