+++ lib/src/models/wagon_wheel_leg_off_label_properties.dart
@@ -29,4 +29,3 @@ class WagonWheelLegOffLabelProperties {
     this.fontWeight,
   });
 }
-